Handle multiple Python functions
Since we can't save the strings as members of the PythonFunc structs and
still have them Copy-able, we store indices and keep the strings in a
static hash map.  It turns out that module-level static objects are
tough in Rust, so I've found a macro package to help us out.  Ideally,
we'd read in the functions from an external resource, and use this to
assemble to api javascript, but for now there's a bit of duplication.